[SDK] [Upg]: Change limit to unsigned int

+++ sdk/nyris.sdk/Network/API/XOptions/CategoryPredictionOptions.cs
@@ -9,7 +9,7 @@ namespace Nyris.Sdk.Network.API.XOptions
             Reset();
         }
 
-        public int Threshold { get; set; }
+        public float Threshold { get; set; }
 
         public int Limit { get; set; }
